時計コンポーネント分け
